Refactor api using include_router to split up api modules.
