+++ survos2/frontend/plugins/analyzers/label_analysis.py
@@ -1,25 +1,19 @@
 import numpy as np
 from loguru import logger
-from matplotlib.backends.backend_agg import FigureCanvasAgg
-from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
-from matplotlib.figure import Figure
-from napari.qt.progress import progress
-from survos2.entity.cluster.cluster_plotting import cluster_scatter, image_grid, plot_clustered_img
+
 
 from survos2.frontend.control import Launcher
 from survos2.frontend.components.base import (
     HWidgets,
-    Slider,
     PushButton,
     Card,
     SimpleComboBox,
-    LineEdit,
-    CheckBox,
+    LineEdit
 )
 
 from survos2.model import DataModel
 
-from survos2.frontend.plugins.analyzers.base import AnalyzerCardBase, MplCanvas
+from survos2.frontend.plugins.analyzers.base import AnalyzerCardBase
 from survos2.frontend.plugins.analyzers.constants import feature_names
 
 
@@ -192,40 +186,6 @@ class LabelAnalyzer(AnalyzerCardBase):
             self.display_splitter_results(result_features)
             self.display_splitter_plot(feature_arrays, titles=feature_titles, vert_line_at=None)
 
-
-class RemoveMaskedObjects(AnalyzerCardBase):
-    def __init__(self, analyzer_id, analyzer_name, analyzer_type, parent=None):
-        super().__init__(
-            analyzer_name=analyzer_name,
-            analyzer_id=analyzer_id,
-            analyzer_type=analyzer_type,
-            title=analyzer_name,
-            collapsible=True,
-            removable=True,
-            editable=True,
-            parent=parent,
-        )
-
-    def setup(self):
-        self._add_feature_source()
-        self._add_objects_source()
-        self.invert_checkbox = CheckBox(checked=True)
-        self.add_row(HWidgets("Invert:", self.invert_checkbox, stretch=0))
-        self.load_as_objects_btn = PushButton("Load as Objects")
-        self.additional_buttons.append(self.load_as_objects_btn)
-        self.load_as_objects_btn.clicked.connect(self.load_as_objects)
-
-    def calculate(self):
-        dst = DataModel.g.dataset_uri(self.analyzer_id, group="analyzer")
-        src = DataModel.g.dataset_uri(self.feature_source.value())
-        all_params = dict(src=src, dst=dst, modal=False)
-        all_params["feature_id"] = str(self.feature_source.value())
-        all_params["object_id"] = str(self.objects_source.value())
-        all_params["invert"] = self.invert_checkbox.value()
-
-        result = Launcher.g.run("analyzer", "remove_masked_objects", **all_params)
-        logger.debug(f"remove_masked_objects result table {len(result)}")
-        self.display_component_results(result)
 
 
 class FindConnectedComponents(AnalyzerCardBase):
